Independent XmlParserHandler
